pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "AI-powered news aggregator that scrapes YouTube channels and blogs, generates daily digests using LLM, and sends email summaries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docling>=2.64.0",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "lxml_html_clean>=0.1.1",
    "newspaper3k>=0.2.8",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "redis>=5.2.1",
    "requests>=2.32.5",
    "rq>=2.0.1",
    "sentence-transformers>=3.3.1",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.0.0",
    "youtube-transcript-api>=1.2.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "agent"]

